Removes IC2 Maven, and updates IC2 API

+++ src/main/java/ic2/api/energy/tile/IEnergyConductor.java
@@ -4,7 +4,7 @@ package ic2.api.energy.tile;
  * Tile entities which conduct energy pulses without buffering (mostly cables) have to implement this
  * interface.
  * 
- * See ic2/buildcraft.cofh.ic2.api/energy/usage.txt for an overall description of the energy net buildcraft.cofh.ic2.api.
+ * See ic2/api/energy/usage.txt for an overall description of the energy net api.
  */
 public interface IEnergyConductor extends IEnergyAcceptor, IEnergyEmitter {
 	/**
@@ -19,7 +19,7 @@ public interface IEnergyConductor extends IEnergyAcceptor, IEnergyEmitter {
 	 * 
 	 * @return Insulation energy absorption in EU
 	 */
-	int getInsulationEnergyAbsorption();
+	double getInsulationEnergyAbsorption();
 
 	/**
 	 * Amount of energy the insulation will handle before it is destroyed.
@@ -27,14 +27,14 @@ public interface IEnergyConductor extends IEnergyAcceptor, IEnergyEmitter {
 	 *
 	 * @return Insulation-destroying energy in EU
 	 */
-	int getInsulationBreakdownEnergy();
+	double getInsulationBreakdownEnergy();
 
 	/**
 	 * Amount of energy the conductor will handle before it melts.
 	 * 
 	 * @return Conductor-destroying energy in EU
 	 */
-	int getConductorBreakdownEnergy();
+	double getConductorBreakdownEnergy();
 
 	/**
 	 * Remove the conductor's insulation if the insulation breakdown energy was exceeded.
